Python/requirements.txt
certifi==2018.4.16
chardet==3.0.4
idna==2.6
lxml==4.2.1
matlabruntimeforpython===R2017a
numpy==1.14.2
paho-mqtt==1.3.1
pandas==0.22.0
pandas-datareader==0.6.0
pkg-resources==0.0.0
python-dateutil==2.7.2
pytz==2018.4
requests==2.18.4
requests-file==1.4.3
requests-ftp==0.3.1
six==1.11.0
urllib3==1.22
wrapt==1.10.11
